# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "EEducation API service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1234",
    "basePath": "/",
    "paths": {
        "/admin/lessons": {
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Set new lesson",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Set new lesson \\\\ ADMIN ONLY",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lesson_service.SetLessonReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "409": {
                        "description": "Lesson already exists"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
...
        "/lessons/grades": {
            "get": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Get all student lesson grades by student id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get student's lesson grades",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request",
                        "name": "lesson",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/grade.Grade"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
        "/lessons/grades/all": {
            "get": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Get all student lesson grades by student id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all student's lesson grades",
                "parameters": [
                    {
                        "type": "string",
                        "description": "method could call without any parameters",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/grade.Grade"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
...
        "/teacher/lessons/grades": {
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Set grade by student id \u0026\u0026 lesson",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Set grade",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/grade.SetGradeReq"
                        }
                    },
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/grade.SetGradeReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Grade created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "409": {
                        "description": "Grade already exists"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Delete grade",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete grade",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/grade.DeleteGradeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grade deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "Grade not found"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Change grade",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Change grade",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/grade.ChangeGradeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grade changed"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "404": {
                        "description": "Grade not found"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
...
        "/teacher/lessons/term-grades": {
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Set term grade",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Set term grade",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/grade.SetTermGradeReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Grade created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "406": {
                        "description": "Not enough grades"
                    },
                    "409": {
                        "description": "Grade already exists"
                    },
                    "500": {
                        "description": "Internal"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "grade.ChangeGradeReq": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "grade": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                },
                "student_id": {
                    "type": "integer"
                }
            }
        },
        "grade.DeleteGradeReq": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "lesson_name": {
                    "type": "string"
                },
                "student_id": {
                    "type": "integer"
                }
            }
        },
        "grade.Grade": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "lessonName": {
                    "type": "string"
                },
                "studentId": {
                    "type": "integer"
                }
            }
        },
        "grade.SetGradeReq": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "integer"
                },
                "grade": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                },
                "student_id": {
                    "type": "integer"
                }
            }
        },
        "grade.SetTermGradeReq": {
            "type": "object",
            "properties": {
                "lesson_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "lesson_service.SetLessonReq": {
            "type": "object",
            "properties": {
                "class_name": {
                    "type": "string"
                },
                "lesson_name": {
                    "type": "string"
                },
                "teacher_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "TokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}